platform-hacksaw: Properly deallocate devices.
When a device is deleted, its memory needs to be deallocatd. The
Platform class is using a std::map, which will happily call the
destructor of objects when items are removed with erase(), but until
this patch, the map held straight pointers instead of destructable
objects.

This patch changes the map to contain std::unique_ptr<Device> objects,
and Function* pointers. This works (and will continue to work) so long
as all Device objects own and deallocate their respective Functions (and
so long as these owned Functions are the only Function* values placed in
the map). Likewise, only Devices allocated with "new" should be placed
in the map.