msp430: Fix the cross-compile.
So. The toolchain went and silently ate itself again. The reason my
firmware all mysteriously stopped working is because the toolchain
defaults to MSP430X code now, and those opcodes get used in the function
prologue/epilogues.

To fix this, I had to add the -mmcu=msp430 flag, AND manually add three
-L directories to grab the 430 libraries rather than the 430x ones
(because the builtin specs don't seem to be smart enough to do the WHOLE
job when you ask for msp430 ISA). Furthermore, I had to make sure that
the linker script in the 430 directory had the same patches applied to
it as the script in the 430X location (e.g., that .text got planted in
rom, not ram). Remember kids: the default scripts that come with the GCC
MSP430 cross-compiler are not actually useful for any hardware. You'll
need an extra package for per-chip memory/peripheral maps, and to change
the bloody default link script to stuff data in sane locations.

Verified at the very least that the USB COM demo is printing its little
"hello world" and echoing characters back. Hopefully this will unblock
development on the e-pawn-receiver firmware.